% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{downloadData}
\alias{downloadData}
\title{Download project data from OSF repository.}
\usage{
downloadData(
  folder = "data",
  unzip = TRUE,
  removezips = TRUE,
  checkfiles = TRUE,
  groups = "all",
  sections = c("aligned", "rotated"),
  overwrite = TRUE
)
}
\arguments{
\item{folder}{Directory on the working directory to store the downloaded data.}

\item{removezips}{(Boolean) Remove downloaded zipfiles after unzipping.}

\item{checkfiles}{(Boolean) Check if all files are there after unzipping.}

\item{groups}{Vector of groups to download data for ('all' for all groups).
Possible values: 'control', 'instructed', 'control60', 'instructed60',
'cursorjump', 'handview', 'older_control', 'older_instructed'
'EDS', 'EDSmatch', 'org_control', 'org_instructed', 'org_control60', and
'org_instructed60'. The last 4 groups are incomplete pilot studies.
Default: 'all'.}

\item{sections}{Vector of sections of data to download for each of the groups
('all' for all sections). Possible values: 'aligned', 'rotated', and
'localization-reaches'. Default: c('aligned', 'rotated')}
}
\value{
Nothing
}
\description{
Use this function to download the data for the project from the
OSF repository: https://osf.io/dhk3u/
}
\details{
Not yet.
}
